refactored entities and datacontext a bit
