+++ GarbageCan/Data/Entities/Boosters/EntityActiveBooster.cs
@@ -2,7 +2,7 @@
 using System.ComponentModel.DataAnnotations;
 using System.ComponentModel.DataAnnotations.Schema;
 
-namespace GarbageCan.Data.Entities
+namespace GarbageCan.Data.Entities.Boosters
 {
 	[Table("xp_active_boosters")]
 	public class EntityActiveBooster
