+++ GarbageCan/Data/Entities/XP/EntityUser.cs
@@ -2,7 +2,7 @@
 using System.ComponentModel.DataAnnotations;
 using System.ComponentModel.DataAnnotations.Schema;
 
-namespace GarbageCan.Data.Entities
+namespace GarbageCan.Data.Entities.XP
 {
 	[Table("xp_users")]
 	public class EntityUser
